requirements.txt
flask
mysql-connector-python
pyngrok
requests
llama-cpp-python
PyMuPDF
python-docx
beautifulsoup4
python-pptx
Pillow
pytesseract
openpyxl